+++ backend/src/mistral/mod.rs
@@ -1,3 +1,2 @@
 pub mod infer;
 pub mod types;
-pub mod utils;
