+++ backend/src/mistral/types/mod.rs
@@ -1,2 +1,3 @@
-pub mod config;
+pub mod inference_args;
+pub mod load_model;
 pub mod text_generation;
