+++ backend/src/server/rest/mod.rs
@@ -1,0 +1,1 @@
+pub mod inference;
