+++ backend/src/server/websocket/utils/mod.rs
@@ -1,0 +1,2 @@
+pub mod handler;
+pub mod mistral;
