--- backend/src/websockets/mod.rs
@@ -1,3 +1,0 @@
-pub mod server;
-pub mod utils;
-pub mod types;
